reset token migration

+++ database/migrations/2023_08_10_133504_add_reset_token_to_users.php
@@ -13,9 +13,13 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('reset_token');
-        });
+
+        if (Schema::hasColumn('users', 'reset_token')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('reset_token');
+            });
+        }
+
         Schema::table('users', function (Blueprint $table) {
             $table->string('reset_token')->nullable();
         });
